Register Users and Groups menu items from the viewset
